+++ packages/components/pages/f-contact-preferences/src/components/_tests/ContactPreferences.test.js
@@ -51,19 +51,16 @@ const createStore = ({ state, actions }) => new Vuex.Store({
     }
 });
 
-const mountContactPreferences = ({
+const mountContactPreferences = async ({
     actions = storeActions,
     state = storeState,
     mocks = sutMocks,
     propsData = sutProps,
-    data = dataDefaults,
-    initialiseOverride = null
+    data = dataDefaults
 } = {}) => {
     const store = createStore({ state, actions });
 
-    if (initialiseOverride) {
-        ContactPreferences.methods.initialise = initialiseOverride;
-    }
+    initialiseSpy = jest.spyOn(ContactPreferences.methods, 'initialise');
 
     const mock = shallowMount(ContactPreferences, {
         i18n,
@@ -73,6 +70,7 @@ const mountContactPreferences = ({
         store,
         mocks
     });
+    await mock.vm.$nextTick();
 
     return mock;
 };
@@ -102,14 +100,13 @@ describe('ContactPreferences Component', () => {
     });
 
     afterEach(() => {
-        initialiseSpy = null;
         jest.clearAllMocks();
     });
 
     describe('when creating the component', () => {
-        it('should register the Store Module', () => {
+        it('should register the Store Module', async () => {
             // Arrange & Act
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
 
             // Assert
             expect(wrapper.vm.$store.state.fContactPreferencesModule).toBeDefined();
@@ -117,9 +114,9 @@ describe('ContactPreferences Component', () => {
     });
 
     describe('when mounting the component and calling `initialise`', () => {
-        it('should call the load action with the correct parameters', () => {
+        it('should call the load action with the correct parameters', async () => {
             // Arrange & Act
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
 
             // Assert
             expect(storeActions.loadPreferences).toHaveBeenCalledWith(expect.any(Object), {
@@ -132,22 +129,22 @@ describe('ContactPreferences Component', () => {
             });
         });
 
-        it('should set shouldShowErrorPage flag to true if an error occurs', () => {
+        it('should set shouldShowErrorPage flag to true if an error occurs', async () => {
             // Arrange & Act
             const errorActions = {
                 loadPreferences: jest.fn().mockImplementationOnce(() => {
                     throw new Error('some-error');
                 })
             };
-            wrapper = mountContactPreferences({ actions: errorActions });
+            wrapper = await mountContactPreferences({ actions: errorActions });
 
             // Assert
             expect(wrapper.vm.shouldShowErrorPage).toEqual(true);
         });
 
-        it('should not show the error card if no errors', () => {
+        it('should not show the error card if no errors', async () => {
             // Arrange & Act
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             const element = wrapper.find('[data-test-id="contact-preferences-error-card"]');
 
             // Assert
@@ -156,7 +153,7 @@ describe('ContactPreferences Component', () => {
 
         it('should show the error card if shouldShowErrorPage is true', async () => {
             // Arrange & Act
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             await wrapper.setData({ shouldShowErrorPage: true });
             const element = wrapper.find('[data-test-id="contact-preferences-error-card"]');
 
@@ -166,11 +163,11 @@ describe('ContactPreferences Component', () => {
 
         it('should set the `isFormDirty` flag to false', async () => {
             // Arrange
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             await wrapper.setData({ isFormDirty: true });
 
             // Act - re-mount the page to prove assertion
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
 
             // Assert
             expect(wrapper.vm.isFormDirty).toEqual(false);
@@ -178,8 +175,7 @@ describe('ContactPreferences Component', () => {
 
         it('should log an info log', async () => {
             // Arrange & Act
-            wrapper = mountContactPreferences();
-            await wrapper.vm.$nextTick();
+            wrapper = await mountContactPreferences();
 
             // Assert
             expect(logMocks.info).toHaveBeenCalledTimes(1);
@@ -195,20 +191,18 @@ describe('ContactPreferences Component', () => {
             };
 
             // Act
-            wrapper = mountContactPreferences({ actions: errorActions });
-            await wrapper.vm.$nextTick();
+            wrapper = await mountContactPreferences({ actions: errorActions });
 
             // Assert
             expect(logMocks.error).toHaveBeenCalledTimes(1);
         });
 
-        it('should not call initialise method if the authorisation has not completed', () => {
+        it('should not call initialise method if the authorisation has not completed', async () => {
             // Arrange
-            initialiseSpy = jest.fn();
             sutProps = { ...sutProps, isAuthFinished: false };
 
             // Act
-            wrapper = mountContactPreferences({ initialiseOverride: initialiseSpy });
+            wrapper = await mountContactPreferences();
 
             // Assert
             expect(initialiseSpy).not.toHaveBeenCalled();
@@ -216,11 +210,10 @@ describe('ContactPreferences Component', () => {
 
         it('should only call initialise method once the authorisation has completed', async () => {
             // Arrange
-            initialiseSpy = jest.fn();
             sutProps = { ...sutProps, isAuthFinished: false };
 
             // Act
-            wrapper = mountContactPreferences({ initialiseOverride: initialiseSpy });
+            wrapper = await mountContactPreferences();
 
             // Assert
             expect(initialiseSpy).not.toHaveBeenCalled();
@@ -239,10 +232,10 @@ describe('ContactPreferences Component', () => {
             ['news', 'email', false, true],
             ['news', 'sms', true, false],
             ['news', 'sms', false, true]
-        ])('should call the Mutation correctly for the %s preferences %s checkbox with %s if previously %s', (preferencesKey, preferenceName, setValue, previousValue) => {
+        ])('should call the Mutation correctly for the %s preferences %s checkbox with %s if previously %s', async (preferencesKey, preferenceName, setValue, previousValue) => {
             // Arrange
             storeState.preferences.find(e => e.key === preferencesKey)[`${preferenceName}Value`] = previousValue;
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             const input = wrapper.find(`[data-test-id="contact-preferences-${preferencesKey}-${preferenceName}-checkbox"]`);
 
             // Act
@@ -259,9 +252,9 @@ describe('ContactPreferences Component', () => {
             );
         });
 
-        it('should set the `isFormDirty` flag to true', () => {
+        it('should set the `isFormDirty` flag to true', async () => {
             // Act
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             const input = wrapper.find('[data-test-id="contact-preferences-news-email-checkbox"]');
 
             // Act
@@ -275,11 +268,11 @@ describe('ContactPreferences Component', () => {
     describe('when clicking the save preferences button', () => {
         it('should call the save action with the correct parameters', async () => {
             // Arrange
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             await wrapper.setData({ isFormDirty: true });
 
             // Act
-            wrapper.vm.onFormSubmit();
+            await wrapper.vm.onFormSubmit();
 
             // Assert
             expect(storeActions.savePreferences).toHaveBeenCalledWith(expect.any(Object), {
@@ -294,8 +287,9 @@ describe('ContactPreferences Component', () => {
 
         it('should log an info log', async () => {
             // Arrange
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             await wrapper.setData({ isFormDirty: true });
+            logMocks.info.mockClear(); // initialise has already logged info once
 
             // Act
             await wrapper.vm.onFormSubmit();
@@ -306,7 +300,7 @@ describe('ContactPreferences Component', () => {
 
         it('should not call the save action if no changes', async () => {
             // Arrange
-            wrapper = mountContactPreferences();
+            wrapper = await mountContactPreferences();
             await wrapper.setData({ isFormDirty: false });
 
             // Act
@@ -324,7 +318,7 @@ describe('ContactPreferences Component', () => {
                     throw new Error('some-error');
                 })
             };
-            wrapper = mountContactPreferences({ actions: errorActions });
+            wrapper = await mountContactPreferences({ actions: errorActions });
             await wrapper.setData({ isFormDirty: true }); // Allow the data to be submitted
 
             // Act
@@ -342,7 +336,7 @@ describe('ContactPreferences Component', () => {
                     throw new Error('some-error');
                 })
             };
-            wrapper = mountContactPreferences({ actions: errorActions });
+            wrapper = await mountContactPreferences({ actions: errorActions });
             await wrapper.setData({ isFormDirty: true });
 
             // Act